rewrite using FnMut -> impl Future
This is a shame, because it's much more restrictive than AsyncFn in
terms of what you can borrow, but it seems like it's necessary to avoid
blowing up compile times until return type notation is stable. (It's
also what `StreamsExt` does throughout.) We can at least retain Fn/FnMut
in the for_each calls at the end, since that doesn't need to be stored.